PR feedback. Further verification of P2P messaging required before merging.
