+++ applications/workers/release/rpc-worker/src/e2eTest/kotlin/net/corda/applications/workers/rpc/MultiClusterDynamicNetworkTest.kt
@@ -1,43 +1,37 @@
 package net.corda.applications.workers.rpc
 
 import net.corda.applications.workers.rpc.http.TestToolkitProperty
-import net.corda.applications.workers.rpc.utils.ALICE_CLUSTER_NS
-import net.corda.applications.workers.rpc.utils.BOB_CLUSTER_NS
 import net.corda.applications.workers.rpc.utils.ClusterTestData
 import net.corda.applications.workers.rpc.utils.HSM_CAT_LEDGER
 import net.corda.applications.workers.rpc.utils.HSM_CAT_SESSION
 import net.corda.applications.workers.rpc.utils.HSM_CAT_TLS
-import net.corda.applications.workers.rpc.utils.MGM_CLUSTER_NS
 import net.corda.applications.workers.rpc.utils.MemberTestData
-import net.corda.applications.workers.rpc.utils.P2P_GATEWAY
 import net.corda.applications.workers.rpc.utils.P2P_TENANT_ID
-import net.corda.applications.workers.rpc.utils.RPC_PORT
-import net.corda.applications.workers.rpc.utils.RPC_WORKER
 import net.corda.applications.workers.rpc.utils.assertMemberInMemberList
 import net.corda.applications.workers.rpc.utils.assertOnlyMgmIsInMemberList
+import net.corda.applications.workers.rpc.utils.assertP2pConnectivity
 import net.corda.applications.workers.rpc.utils.assignSoftHsm
 import net.corda.applications.workers.rpc.utils.createMGMGroupPolicyJson
+import net.corda.applications.workers.rpc.utils.createMemberRegistrationContext
+import net.corda.applications.workers.rpc.utils.createMgmRegistrationContext
 import net.corda.applications.workers.rpc.utils.createVirtualNode
 import net.corda.applications.workers.rpc.utils.disableCLRChecks
-import net.corda.applications.workers.rpc.utils.genGroupPolicy
-import net.corda.applications.workers.rpc.utils.genKeyPair
+import net.corda.applications.workers.rpc.utils.generateGroupPolicy
+import net.corda.applications.workers.rpc.utils.generateKeyPair
 import net.corda.applications.workers.rpc.utils.generateCert
 import net.corda.applications.workers.rpc.utils.generateCsr
 import net.corda.applications.workers.rpc.utils.getCa
-import net.corda.applications.workers.rpc.utils.getMemberRegistrationContext
-import net.corda.applications.workers.rpc.utils.getMgmRegistrationContext
-import net.corda.applications.workers.rpc.utils.groupId
+import net.corda.applications.workers.rpc.utils.getGroupId
 import net.corda.applications.workers.rpc.utils.keyExists
 import net.corda.applications.workers.rpc.utils.lookupMembers
 import net.corda.applications.workers.rpc.utils.name
 import net.corda.applications.workers.rpc.utils.register
 import net.corda.applications.workers.rpc.utils.setUpNetworkIdentity
 import net.corda.applications.workers.rpc.utils.status
-import net.corda.applications.workers.rpc.utils.tmpPath
-import net.corda.applications.workers.rpc.utils.toByteArray
 import net.corda.applications.workers.rpc.utils.uploadCpi
 import net.corda.applications.workers.rpc.utils.uploadTlsCertificate
 import net.corda.crypto.test.certificates.generation.toPem
+import net.corda.data.identity.HoldingIdentity
 import net.corda.membership.httprpc.v1.types.response.RpcMemberInfo
 import net.corda.test.util.eventually
 import net.corda.v5.base.util.seconds
@@ -45,7 +39,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.Test
-import java.io.File
 
 /**
  * When running this test locally, three clusters will most likely be too much for your laptop.
@@ -82,27 +75,35 @@ import java.io.File
 )
 class MultiClusterDynamicNetworkTest {
 
-    private val mgmRpcHost = "$RPC_WORKER.$MGM_CLUSTER_NS"
-    private val mgmP2pHost = "$P2P_GATEWAY.$MGM_CLUSTER_NS"
-    private val mgmTestToolkit by TestToolkitProperty(mgmRpcHost, RPC_PORT)
+    private val mgmRpcHost = System.getProperty("e2eClusterCRpcHost")
+    private val mgmRpcPort = System.getProperty("e2eClusterCRpcPort").toInt()
+    private val mgmP2pHost = System.getProperty("e2eClusterCP2pHost")
+    private val mgmP2pPort = System.getProperty("e2eClusterCP2pPort").toInt()
+    private val mgmTestToolkit by TestToolkitProperty(mgmRpcHost, mgmRpcPort)
 
-    private val aliceRpcHost = "$RPC_WORKER.$ALICE_CLUSTER_NS"
-    private val aliceP2pHost = "$P2P_GATEWAY.$ALICE_CLUSTER_NS"
-    private val aliceTestToolkit by TestToolkitProperty(aliceRpcHost, RPC_PORT)
+    private val aliceRpcHost = System.getProperty("e2eClusterARpcHost")
+    private val aliceRpcPort = System.getProperty("e2eClusterARpcPort").toInt()
+    private val aliceP2pHost = System.getProperty("e2eClusterAP2pHost")
+    private val aliceP2pPort = System.getProperty("e2eClusterAP2pPort").toInt()
+    private val aliceTestToolkit by TestToolkitProperty(aliceRpcHost, aliceRpcPort)
 
-    private val bobRpcHost = "$RPC_WORKER.$BOB_CLUSTER_NS"
-    private val bobP2pHost = "$P2P_GATEWAY.$BOB_CLUSTER_NS"
-    private val bobTestToolkit by TestToolkitProperty(bobRpcHost, RPC_PORT)
+    private val bobRpcHost = System.getProperty("e2eClusterBRpcHost")
+    private val bobRpcPort = System.getProperty("e2eClusterBRpcPort").toInt()
+    private val bobP2pHost = System.getProperty("e2eClusterBP2pHost")
+    private val bobP2pPort = System.getProperty("e2eClusterBP2pPort").toInt()
+    private val bobTestToolkit by TestToolkitProperty(bobRpcHost, bobRpcPort)
 
     private val mgmCluster = ClusterTestData(
         mgmTestToolkit,
         mgmP2pHost,
+        mgmP2pPort,
         listOf(MemberTestData("O=Mgm, L=London, C=GB, OU=${mgmTestToolkit.uniqueName}"))
     )
 
     private val aliceCluster = ClusterTestData(
         aliceTestToolkit,
         aliceP2pHost,
+        aliceP2pPort,
         listOf(
             MemberTestData("O=Alice, L=London, C=GB, OU=${aliceTestToolkit.uniqueName}"),
         )
@@ -110,6 +111,7 @@ class MultiClusterDynamicNetworkTest {
     private val bobCluster = ClusterTestData(
         bobTestToolkit,
         bobP2pHost,
+        bobP2pPort,
         listOf(
             MemberTestData("O=Bob, L=London, C=GB, OU=${bobTestToolkit.uniqueName}"),
         )
@@ -121,6 +123,33 @@ class MultiClusterDynamicNetworkTest {
 
     @Test
     fun `Create mgm and allow members to join the group`() {
+        onboardMultiClusterGroup()
+    }
+
+    @Disabled("Is disabled and can be run manually until CORE-6079 is complete. At that point this can be " +
+        "merged into the above test.")
+    @Test
+    fun `Onboard group across clusters and check p2p connectivity`() {
+        val groupId = onboardMultiClusterGroup()
+
+        assertP2pConnectivity(
+            HoldingIdentity(
+                aliceCluster.members.first().name,
+                groupId
+            ),
+            HoldingIdentity(
+                bobCluster.members.first().name,
+                groupId
+            ),
+            aliceCluster.kafkaTestToolkit,
+            bobCluster.kafkaTestToolkit
+        )
+    }
+
+    /**
+     * Onboard group and return group ID.
+     */
+    private fun onboardMultiClusterGroup(): String {
         // For the purposes of this test, the MGM cluster is
         // expected to have only one MGM (in reality there can be more on a cluster).
         assertThat(mgmCluster.members).hasSize(1)
@@ -131,15 +160,14 @@ class MultiClusterDynamicNetworkTest {
         val cpiChecksum = mgmCluster.uploadCpi(createMGMGroupPolicyJson(), true)
         val mgmHoldingId = mgmCluster.createVirtualNode(mgm, cpiChecksum)
         holdingIds[mgm.name] = mgmHoldingId
-        println("MGM HoldingIdentity: $mgmHoldingId")
 
         mgmCluster.assignSoftHsm(mgmHoldingId, HSM_CAT_SESSION)
 
-        val mgmSessionKeyId = mgmCluster.genKeyPair(mgmHoldingId, HSM_CAT_SESSION)
+        val mgmSessionKeyId = mgmCluster.generateKeyPair(mgmHoldingId, HSM_CAT_SESSION)
 
         mgmCluster.register(
             mgmHoldingId,
-            getMgmRegistrationContext(
+            createMgmRegistrationContext(
                 tlsTrustRoot = ca.caCertificate.toPem(),
                 sessionKeyId = mgmSessionKeyId,
                 p2pUrl = mgmCluster.p2pUrl
@@ -149,7 +177,7 @@ class MultiClusterDynamicNetworkTest {
         mgmCluster.assertOnlyMgmIsInMemberList(mgmHoldingId, mgm.name)
 
         if (!mgmCluster.keyExists(P2P_TENANT_ID, HSM_CAT_TLS)) {
-            val mgmTlsKeyId = mgmCluster.genKeyPair(P2P_TENANT_ID, HSM_CAT_TLS)
+            val mgmTlsKeyId = mgmCluster.generateKeyPair(P2P_TENANT_ID, HSM_CAT_TLS)
             val mgmTlsCsr = mgmCluster.generateCsr(mgm, mgmTlsKeyId)
             val mgmTlsCert = ca.generateCert(mgmTlsCsr)
             mgmCluster.uploadTlsCertificate(mgmTlsCert)
@@ -160,29 +188,28 @@ class MultiClusterDynamicNetworkTest {
             mgmSessionKeyId
         )
 
-        val memberGroupPolicy = mgmCluster.genGroupPolicy(mgmHoldingId)
+        val memberGroupPolicy = mgmCluster.generateGroupPolicy(mgmHoldingId)
 
         clusters.forEach { cordaCluster ->
             cordaCluster.disableCLRChecks()
-            val memberCpiChecksum = cordaCluster.uploadCpi(toByteArray(memberGroupPolicy))
+            val memberCpiChecksum = cordaCluster.uploadCpi(memberGroupPolicy.toByteArray())
 
             cordaCluster.members.forEach { member ->
                 val memberHoldingId = cordaCluster.createVirtualNode(member, memberCpiChecksum)
                 holdingIds[member.name] = memberHoldingId
-                println("${member.name} holding ID: $memberHoldingId")
 
                 cordaCluster.assignSoftHsm(memberHoldingId, HSM_CAT_SESSION)
                 cordaCluster.assignSoftHsm(memberHoldingId, HSM_CAT_LEDGER)
 
                 if (!cordaCluster.keyExists(P2P_TENANT_ID, HSM_CAT_TLS)) {
-                    val memberTlsKeyId = cordaCluster.genKeyPair(P2P_TENANT_ID, HSM_CAT_TLS)
+                    val memberTlsKeyId = cordaCluster.generateKeyPair(P2P_TENANT_ID, HSM_CAT_TLS)
                     val memberTlsCsr = cordaCluster.generateCsr(member, memberTlsKeyId)
                     val memberTlsCert = ca.generateCert(memberTlsCsr)
                     cordaCluster.uploadTlsCertificate(memberTlsCert)
                 }
 
-                val memberSessionKeyId = cordaCluster.genKeyPair(memberHoldingId, HSM_CAT_SESSION)
-                val memberLedgerKeyId = cordaCluster.genKeyPair(memberHoldingId, HSM_CAT_LEDGER)
+                val memberSessionKeyId = cordaCluster.generateKeyPair(memberHoldingId, HSM_CAT_SESSION)
+                val memberLedgerKeyId = cordaCluster.generateKeyPair(memberHoldingId, HSM_CAT_LEDGER)
 
                 cordaCluster.setUpNetworkIdentity(
                     memberHoldingId,
@@ -192,7 +219,7 @@ class MultiClusterDynamicNetworkTest {
                 cordaCluster.assertOnlyMgmIsInMemberList(memberHoldingId, mgm.name)
                 cordaCluster.register(
                     memberHoldingId,
-                    getMemberRegistrationContext(
+                    createMemberRegistrationContext(
                         cordaCluster,
                         memberSessionKeyId,
                         memberLedgerKeyId
@@ -217,8 +244,8 @@ class MultiClusterDynamicNetworkTest {
                 assertNotNull(holdingId)
 
                 eventually(
-                    waitBetween = 2.seconds,
-                    duration = 10.seconds
+                    waitBetween = 3.seconds,
+                    duration = 60.seconds
                 ) {
                     val memberList = cordaCluster.lookupMembers(holdingId!!)
 
@@ -246,58 +273,6 @@ class MultiClusterDynamicNetworkTest {
                 }
             }
         }
-
-        /**
-         * This is a temporary addition to the test. This function prints some commands to the console to assist in
-         * manually confirming that the P2P messages can be sent between members which registered.
-         * This is temporary until it is possible to access kafka directly to confirm.
-         */
-        printAppSimulatorConfig(
-            clusterMembers[clusters[0]]!!.first(),
-            clusterMembers[clusters[1]]!!.first()
-        )
-    }
-
-    private fun printAppSimulatorConfig(
-        source: RpcMemberInfo,
-        destination: RpcMemberInfo
-    ) {
-        val senderConfig = """{
-                parallelClients: 1,
-                simulatorMode: "SENDER",
-                loadGenerationParams: {
-                    peerX500Name: "${destination.name}",
-                    peerGroupId: "${destination.groupId}",
-                    ourX500Name: "${source.name}",
-                    ourGroupId: "${source.groupId}",
-                    loadGenerationType: "ONE_OFF",
-                    batchSize: 1,
-                    totalNumberOfMessages: 1,
-                    interBatchDelay: 0ms,
-                    messageSizeBytes: 10000
-                }
-            }""".trimIndent()
-        val configFileName = "simulator_sender.conf"
-        val buildDir = File(tmpPath)
-        val simulatorSenderConfig = File(buildDir, configFileName).apply {
-            writeText(senderConfig)
-        }
-
-        val caLocation = "${buildDir.absolutePath}${File.separator}ca.crt"
-
-        println(
-            """
-                The following steps allow for manual verification of connectivity between clusters. View the kafka topic to observe messages being sent across clusters.
-                
-                To test connection get the kafka trust store by running:
-                
-                kubectl get secret prereqs-kafka-0-tls -o go-template='{{ index .data "ca.crt" }}' -n $ALICE_CLUSTER_NS | base64 -D > $caLocation
-                
-                Then run the following with that truststore (Ensure running where corda-app-simulator-5.0.0.0-SNAPSHOT.jar is reachable):
-                
-                java -jar corda-app-simulator-5.0.0.0-SNAPSHOT.jar -mbootstrap.servers=prereqs-kafka.$ALICE_CLUSTER_NS:9092 -msecurity.protocol=SSL -mssl.truststore.location=$caLocation -mssl.truststore.type=PEM  --simulator-config ${simulatorSenderConfig.absolutePath}
-
-        """.trimIndent()
-        )
+        return mgmCluster.getGroupId(mgmHoldingId)
     }
 }
