+++ applications/workers/release/rpc-worker/src/e2eTest/kotlin/net/corda/applications/workers/rpc/utils/ClusterTestUtils.kt
@@ -1,0 +1,214 @@
+package net.corda.applications.workers.rpc.utils
+
+import com.fasterxml.jackson.databind.ObjectMapper
+import net.corda.httprpc.HttpFileUpload
+import net.corda.libs.configuration.endpoints.v1.ConfigRPCOps
+import net.corda.libs.configuration.endpoints.v1.types.ConfigSchemaVersion
+import net.corda.libs.configuration.endpoints.v1.types.UpdateConfigParameters
+import net.corda.libs.cpiupload.endpoints.v1.CpiUploadRPCOps
+import net.corda.libs.virtualnode.endpoints.v1.VirtualNodeRPCOps
+import net.corda.libs.virtualnode.endpoints.v1.types.VirtualNodeRequest
+import net.corda.membership.httprpc.v1.HsmRpcOps
+import net.corda.membership.httprpc.v1.KeysRpcOps
+import net.corda.membership.httprpc.v1.MGMRpcOps
+import net.corda.membership.httprpc.v1.MemberRegistrationRpcOps
+import net.corda.membership.httprpc.v1.NetworkRpcOps
+import net.corda.membership.httprpc.v1.types.request.HostedIdentitySetupRequest
+import net.corda.membership.httprpc.v1.types.request.MemberRegistrationRequest
+import net.corda.test.util.eventually
+import org.assertj.core.api.Assertions.assertThat
+
+const val GATEWAY_CONFIG = "corda.p2p.gateway"
+const val KEY_SCHEME = "CORDA.ECDSA.SECP256R1"
+const val P2P_TENANT_ID = "p2p"
+const val HSM_CAT_SESSION = "SESSION_INIT"
+const val HSM_CAT_LEDGER = "LEDGER"
+const val HSM_CAT_TLS = "TLS"
+
+fun ClusterTestData.uploadCpi(
+    groupPolicy: ByteArray,
+    isMgm: Boolean = false
+) = with(testToolkit) {
+    httpClientFor(CpiUploadRPCOps::class.java).use { client ->
+        with(client.start().proxy) {
+            // Check if MGM CPI was already uploaded in previous run. Current validation only allows one MGM CPI.
+            if (isMgm) {
+                getAllCpis().cpis.firstOrNull {
+                    it.groupPolicy?.contains("CREATE_ID") ?: false
+                }?.let {
+                    return it.cpiFileChecksum
+                }
+            }
+
+            val jar = createEmptyJarWithManifest(groupPolicy)
+            val upload = HttpFileUpload(
+                content = jar.inputStream(),
+                contentType = "application/java-archive",
+                extension = "cpb",
+                fileName = "${uniqueName}.cpb",
+                size = jar.size.toLong(),
+            )
+            val id = cpi(upload).id
+            eventually {
+                val status = status(id)
+                assertThat(status.status).isEqualTo("OK")
+                status.cpiFileChecksum
+            }
+        }
+    }
+}
+
+fun ClusterTestData.createVirtualNode(
+    member: MemberTestData,
+    cpiCheckSum: String
+) = with(testToolkit) {
+    httpClientFor(VirtualNodeRPCOps::class.java)
+        .use { client ->
+            client.start().proxy.createVirtualNode(
+                VirtualNodeRequest(
+                    x500Name = member.name,
+                    cpiFileChecksum = cpiCheckSum,
+                    vaultDdlConnection = null,
+                    vaultDmlConnection = null,
+                    cryptoDdlConnection = null,
+                    cryptoDmlConnection = null,
+                )
+            ).holdingIdentity.shortHash
+        }
+}
+
+fun ClusterTestData.keyExists(
+    tenantId: String,
+    cat: String
+) = with(testToolkit) {
+    httpClientFor(KeysRpcOps::class.java)
+        .use { client ->
+            with(client.start().proxy) {
+                val keyAlias = "$tenantId-$cat"
+                listKeys(
+                    tenantId = tenantId,
+                    skip = 0,
+                    take = 1,
+                    orderBy = "none",
+                    category = cat,
+                    schemeCodeName = null,
+                    alias = keyAlias,
+                    masterKeyAlias = null,
+                    createdAfter = null,
+                    createdBefore = null,
+                    ids = null
+                ).isNotEmpty()
+            }
+        }
+}
+
+fun ClusterTestData.generateKeyPair(
+    tenantId: String,
+    cat: String
+) = with(testToolkit) {
+    httpClientFor(KeysRpcOps::class.java)
+        .use { client ->
+            with(client.start().proxy) {
+                val keyAlias = "$tenantId-$cat"
+                listKeys(
+                    tenantId = tenantId,
+                    skip = 0,
+                    take = 1,
+                    orderBy = "none",
+                    category = cat,
+                    schemeCodeName = null,
+                    alias = keyAlias,
+                    masterKeyAlias = null,
+                    createdAfter = null,
+                    createdBefore = null,
+                    ids = null
+                ).map {
+                    it.value
+                }.firstOrNull()
+                    ?.keyId
+                    ?: generateKeyPair(
+                        tenantId,
+                        keyAlias,
+                        cat,
+                        KEY_SCHEME
+                    )
+            }
+        }
+}
+
+fun ClusterTestData.assignSoftHsm(
+    holdingId: String,
+    cat: String
+) = with(testToolkit) {
+    httpClientFor(HsmRpcOps::class.java)
+        .use { client ->
+            client.start().proxy.assignSoftHsm(holdingId, cat)
+        }
+}
+
+fun ClusterTestData.register(
+    holdingId: String,
+    context: Map<String, String>
+) = with(testToolkit) {
+    httpClientFor(MemberRegistrationRpcOps::class.java)
+        .use { client ->
+            client.start().proxy.startRegistration(
+                holdingId,
+                MemberRegistrationRequest(
+                    action = "requestJoin",
+                    context = context
+                )
+            ).apply {
+                assertThat(registrationStatus).isEqualTo("SUBMITTED")
+            }
+        }
+}
+
+fun ClusterTestData.generateGroupPolicy(
+    holdingId: String
+) = with(testToolkit) {
+    httpClientFor(MGMRpcOps::class.java).use { client ->
+        client.start().proxy.generateGroupPolicy(holdingId)
+    }
+}
+
+fun ClusterTestData.setUpNetworkIdentity(
+    holdingId: String,
+    sessionKeyId: String
+) = with(testToolkit) {
+    httpClientFor(NetworkRpcOps::class.java).use { client ->
+        client.start().proxy.setupHostedIdentities(
+            holdingId,
+            HostedIdentitySetupRequest(
+                TLS_CERT_ALIAS,
+                P2P_TENANT_ID,
+                null,
+                sessionKeyId
+            )
+        )
+    }
+}
+
+fun ClusterTestData.disableCLRChecks() = with(testToolkit) {
+    val sslConfig = "sslConfig"
+    val revocationCheck = "revocationCheck"
+    val mode = "mode"
+    val modeOff = "OFF"
+    httpClientFor(ConfigRPCOps::class.java).use { client ->
+        val proxy = client.start().proxy
+        val configResponse = proxy.get(GATEWAY_CONFIG)
+        val config = ObjectMapper().readTree(
+            configResponse.configWithDefaults
+        )
+        if (modeOff != config[sslConfig][revocationCheck][mode].asText()) {
+            proxy.updateConfig(
+                UpdateConfigParameters(
+                    GATEWAY_CONFIG,
+                    configResponse.version,
+                    "{ \"$sslConfig\": { \"$revocationCheck\": { \"$mode\": \"$modeOff\" }  }  }",
+                    ConfigSchemaVersion(1, 0)
+                )
+            )
+        }
+    }
+}
